Adding more information about line numbers
We now iterate DIA's line number table to retrieve everything in there
Currently this is causing some duplicate entries in the table. I hope to
find a way around that.
We always capture address info for line numbers because it's cheap.
Make sure the src_line id# is actually output...